training/requirements.txt
tensorflow==2.10.0
keras==2.10.0
keras-tuner==1.1.2
scikit-learn==1.0.2
numpy==1.21.5
Pillow==8.4.0
matplotlib==3.5.0
tqdm==4.62.3
